requirements.txt
django==4.1
django-filter
djangorestframework
djangorestframework-simplejwt
pytest-django
pytest-factoryboy